Split files for Invoker classes
