+++ Reflectious/Invoker/GenericMethodInvoker.cs
@@ -1,0 +1,10 @@
+﻿namespace Firestorm
+{
+    public class GenericMethodInvoker
+    {
+        public void Invoke(params object[] args)
+        {
+            
+        }
+    }
+}
